ArabianTotestra: Fix player handicap bonus resources
This code, instead of giving the player a bunch of TECH_NONE bonuses
resources when they turn up the handicap, gives them a bunch of
resources which are useful in Legends of Ancient Arabia.

This is because the LOAA economic model pretty much requires the player
to have camels early in the midgame or economically they are toast; so
even getting the minimal number of handicap resources now always
gives them camels near the player starting plot, even if said starting
plot is in a jungle or grassland.

It’s left as an exercise to the reader why a bunch of camels would be
in the middle of the jungle.