Begin work on ArabianTotestraBig.py
This is ArabianTotestra.py, but the maps are 192x128 instead of 144x96

Note that there is, when rendered as a globe, a visible seam with the
clouds at the international date line (antimeridian). Also note that
the script only has one seed right now, that the seeds are in Python3
compatible 32-bit hex format (as a string), e.g. the one map we have
right now is the string '0x0000_0076'.

Yes, the script right now always generates the same map, but that
will change once we have more suitable seeds.